Introduce ItemBuilder for better test readability
The default Item constructor new Item("aString", aNumber, anotherNumber) was a bit confusing and it was easy to forget which parameter was which.
Now it reads much more clearly as new ItemBuilder().withName("aString").withSellIn("aNumber").withQuality("anotherNumber").build();
Had to keep the name, quality and sellIn values as fields of the builder itself since Item doesn't have a default constructor nor setters or getters.